+++ packages/elements/calendar/examples/2-on-change.js
@@ -2,6 +2,6 @@
 
 import React from 'react';
 import { CalendarStateless } from '../src';
-import { action } from './_';
+import { action } from './utils/_';
 
 export default () => <CalendarStateless onChange={action('change')} />;