+++ packages/fabric/editor-common/src/schema/nodes/image.ts
@@ -11,6 +11,10 @@ export const image: NodeSpec = {
   draggable: true,
   parseDOM: [
     {
+      tag: 'img[src^="data:image/"]',
+      ignore: true,
+    },
+    {
       tag: 'img[src]',
       getAttrs(dom: HTMLElement) {
         return {