+++ packages/elements/empty-state/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @atlaskit/empty-state
 
+## 0.1.2
+
+## 0.1.1
+
 ## 0.1.0
 - [patch] Updates dependency on docs/ to ^1.0.1 [36c7ef8](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/36c7ef8)
 - [minor] Initial release [afbb1e5](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/afbb1e5)